% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos_aprendizaje.R
\name{build_matrix_from_artifact}
\alias{build_matrix_from_artifact}
\title{Construye matriz de predictores desde un artefacto}
\usage{
build_matrix_from_artifact(new_df, artifact)
}
\arguments{
\item{new_df}{data.frame nuevo con columnas originales (antes de dummies).}

\item{artifact}{lista cargada desde \code{guardar_artefacto_modelos()}.}
}
\value{
data.frame X con columnas alineadas a las del modelo.
}
\description{
Construye matriz de predictores desde un artefacto
}
\examples{
# X <- build_matrix_from_artifact(new_df, readRDS("...rds"))
}
\seealso{
Other DiccionariosEmocion: 
\code{\link{analisis_componentes_general}()},
\code{\link{analizar_coherencia_metodos}()},
\code{\link{compute_correlations}()},
\code{\link{compute_kappa}()},
\code{\link{cronbach_alpha}()},
\code{\link{evaluar_con_artefacto}()},
\code{\link{features_canonicas}()},
\code{\link{guardar_artefacto_modelos}()},
\code{\link{normalizar_emociones}()},
\code{\link{predecir_con_artefacto}()},
\code{\link{prep_emotions_data}()},
\code{\link{preparar_features_tokens}()}
}
\concept{DiccionariosEmocion}
