% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetado_sentimiento.R
\name{categorizar_por_percentiles_full}
\alias{categorizar_por_percentiles_full}
\title{Categoriza por percentiles manteniendo longitud (robusto a empates/NA)}
\usage{
categorizar_por_percentiles_full(
  x,
  probs = c(0.25, 0.5, 0.75),
  labels = NULL,
  include_lowest = TRUE,
  right = FALSE
)
}
\arguments{
\item{x}{numeric vector (se mantiene longitud)}

\item{probs}{percentiles deseados (default c(0.25,0.5,0.75))}

\item{labels}{etiquetas a usar; si NULL se autogeneran N-0..N-k}

\item{include_lowest, }{right como en cut()}

\item{0.5}{(auto) TODO: describir par치metro.}

\item{0.75}{(auto) TODO: describir par치metro.}
}
\value{
list(categoria= factor(len(x)), breaks= numeric, probs_usados=numeric)
}
\description{
Descripci칩n (auto). Completar con prop칩sito, entradas y salidas.
}
\seealso{
Other utilidades: 
\code{\link{DividirPorPercentiles}()},
\code{\link{GenerarDataframeT}()},
\code{\link{LimpiaTexto}()},
\code{\link{TabContingencia_MaxMinXFila}()},
\code{\link{TraduceA}()},
\code{\link{UnirExcelITFF_Optimizada}()},
\code{\link{agregar_recuento_filas}()},
\code{\link{agrupar_emociones}()},
\code{\link{ajustar_gradient_boosting}()},
\code{\link{ajustar_red_nn_unificada}()},
\code{\link{ajustar_regresion_multinomial}()},
\code{\link{ajustar_rnn_regresion}()},
\code{\link{ajustar_rnn_regresion_torch}()},
\code{\link{analizar_coherencia_metodos}()},
\code{\link{analizar_comentarios}()},
\code{\link{analizar_comentarios_ReEscalado}()},
\code{\link{analizar_comentarios_ReEscalado_general}()},
\code{\link{analizar_comentarios_ReEscalado_v3}()},
\code{\link{analizar_comentarios_XP}()},
\code{\link{analizar_consistencia_lexicos}()},
\code{\link{analizar_consistencia_lexicos_v2}()},
\code{\link{analizar_correlaciones_metodos}()},
\code{\link{analizar_emociones_texto}()},
\code{\link{analizar_percentiles_columna}()},
\code{\link{analizar_percentiles_columnaBinaria}()},
\code{\link{analizar_tokens_extremos}()},
\code{\link{anotar_tokens_idioma_simple}()},
\code{\link{as_df_safe}()},
\code{\link{build_emotion_indices}()},
\code{\link{calcular_ensamble_ponderado}()},
\code{\link{calcular_puntaje_emocional}()},
\code{\link{calibrar_metodo_referencia}()},
\code{\link{cambiar_nombres_columnas}()},
\code{\link{categorizar_indices}()},
\code{\link{categorizar_por_percentiles}()},
\code{\link{cluster_emotion_spaces}()},
\code{\link{cluster_hclust_cut}()},
\code{\link{cluster_kmeans_auto}()},
\code{\link{configurar_openai_clave}()},
\code{\link{convertir_a_factores_arbol}()},
\code{\link{crear_diccionario_emociones}()},
\code{\link{eliminar_filas_repetidas}()},
\code{\link{eliminarduplicados}()},
\code{\link{emocion_mas_fuerte}()},
\code{\link{estrategia_p1_preparacion}()},
\code{\link{estrategia_p3_ensamble}()},
\code{\link{estrategia_p5_clusters}()},
\code{\link{extraer_features_emocionales}()},
\code{\link{filtrar_diccionario}()},
\code{\link{generar_reporte_html}()},
\code{\link{guardarDataFrameCSV}()},
\code{\link{interseccionDataFrames}()},
\code{\link{leerCSVConNombres}()},
\code{\link{mostrar_diccionario_emociones}()},
\code{\link{p7_exportar_reporte_maestro}()},
\code{\link{pipeline_emociones}()},
\code{\link{predecir_nivel_bin}()},
\code{\link{procesar_emociones_comentarios}()},
\code{\link{run_fa}()},
\code{\link{safe_write_csv}()},
\code{\link{segmentar_consistencia_lexicos}()},
\code{\link{simular_arbol}()},
\code{\link{simular_comparacion_todos}()},
\code{\link{simular_random_forest}()},
\code{\link{simular_red_neuronal}()},
\code{\link{simular_red_nn_unificada}()},
\code{\link{simular_regresion}()},
\code{\link{simular_regresion_multinomial}()},
\code{\link{simular_rnn_regresion}()},
\code{\link{simular_rnn_regresion_torch}()},
\code{\link{sp_save}()},
\code{\link{summarize_emotion_space}()},
\code{\link{vector_a_cadena}()},
\code{\link{verificar_clave_openai}()}
}
\concept{utilidades}
