% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_emociones.R
\name{compute_kappa}
\alias{compute_kappa}
\title{compute_kappa}
\usage{
compute_kappa(df, level_cols)
}
\arguments{
\item{df}{Descripción del parámetro df.}

\item{level_cols}{Descripción del parámetro level_cols.}
}
\value{
Objeto resultante de la función.
}
\description{
(Auto-generado) Documentación roxygen2 mejorada a partir del reporte de auditoría.
}
\examples{
# compute_kappa(df, level_cols)
}
\seealso{
\code{?compute_kappa}

Other DiccionariosEmocion: 
\code{\link{analisis_componentes_general}()},
\code{\link{analizar_coherencia_metodos}()},
\code{\link{build_matrix_from_artifact}()},
\code{\link{compute_correlations}()},
\code{\link{cronbach_alpha}()},
\code{\link{evaluar_con_artefacto}()},
\code{\link{features_canonicas}()},
\code{\link{guardar_artefacto_modelos}()},
\code{\link{normalizar_emociones}()},
\code{\link{predecir_con_artefacto}()},
\code{\link{prep_emotions_data}()},
\code{\link{preparar_features_tokens}()}
}
\concept{DiccionariosEmocion}
