% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos_aprendizaje.R
\name{features_canonicas}
\alias{features_canonicas}
\title{Devuelve el set canónico de features a nivel token}
\usage{
features_canonicas()
}
\value{
character vector con nombres de columnas
}
\description{
Devuelve el set canónico de features a nivel token
}
\examples{
features_canonicas()
}
\seealso{
Other DiccionariosEmocion: 
\code{\link{analisis_componentes_general}()},
\code{\link{analizar_coherencia_metodos}()},
\code{\link{build_matrix_from_artifact}()},
\code{\link{compute_correlations}()},
\code{\link{compute_kappa}()},
\code{\link{cronbach_alpha}()},
\code{\link{evaluar_con_artefacto}()},
\code{\link{guardar_artefacto_modelos}()},
\code{\link{normalizar_emociones}()},
\code{\link{predecir_con_artefacto}()},
\code{\link{prep_emotions_data}()},
\code{\link{preparar_features_tokens}()}
}
\concept{DiccionariosEmocion}
