% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacion_modelos.R
\name{graficar_pr_multiclase}
\alias{graficar_pr_multiclase}
\title{Graficar curvas Precision-Recall para clasificación multiclase}
\usage{
graficar_pr_multiclase(pr_list)
}
\arguments{
\item{pr_list}{Lista de objetos PR por clase, que contengan una curva y AUC.}
}
\value{
Un objeto \code{ggplot2} con las curvas PR por clase. Devuelve \code{NULL} si no hay curvas válidas.
}
\description{
Descripción (auto). Completar con propósito, entradas y salidas.

Esta función genera un gráfico con las curvas Precision-Recall (PR) por clase
a partir de una lista de objetos retornados por \code{PRROC::pr.curve()}.
}
\examples{
graficar_pr_multiclase(resultados_modelo$pr_obj)
}
\seealso{
Other visualizacion: 
\code{\link{CrearGraficoSentimiento}()},
\code{\link{estrategia_p6_estabilidad_visual}()},
\code{\link{graficar_emociones_agrupadas}()},
\code{\link{graficar_importancia_variables}()},
\code{\link{graficar_roc_multiclase}()},
\code{\link{save_plot_list}()},
\code{\link{visualizar_curvas_sentimiento}()},
\code{\link{visualizar_mapa_emocional}()},
\code{\link{visualizar_modelo_polinomial}()}
}
\concept{visualizacion}
