% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplicacion_texto.R
\name{matriz_coocurrencia_nrc}
\alias{matriz_coocurrencia_nrc}
\title{Matriz de coocurrencia NRC}
\usage{
matriz_coocurrencia_nrc(nrc_matrix)
}
\arguments{
\item{nrc_matrix}{data.frame con columnas como emociones NRC y filas como comentarios.}
}
\value{
Matriz de coocurrencia (simétrica).
}
\description{
Calcula la cantidad de veces que las emociones NRC ocurren juntas en comentarios individuales.
}
\examples{
matriz_coocurrencia_nrc(resultados$VectorTextoSentimientosNRC)
cooc <- matriz_coocurrencia_nrc(resultados$VectorTextoSentimientosNRC)
print(cooc)
}
