% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_emociones.R
\name{prep_emotions_data}
\alias{prep_emotions_data}
\title{prep_emotions_data}
\usage{
prep_emotions_data(dfx, method_cols, nrc_cols, id_cols = character(0))

prep_emotions_data(dfx, method_cols, nrc_cols, id_cols = character(0))
}
\arguments{
\item{dfx}{(auto) TODO: describir parámetro.}

\item{method_cols}{(auto) TODO: describir parámetro.}

\item{nrc_cols}{(auto) TODO: describir parámetro.}

\item{id_cols}{(auto) TODO: describir parámetro.}
}
\value{
Objeto resultante de la función.

(auto) Verificar y completar la descripción del valor retornado.
}
\description{
(Auto-generado) Documentación roxygen2 mejorada a partir del reporte de auditoría.

Descripción (auto). Completar con propósito, entradas y salidas.
}
\examples{
# prep_emotions_data(dfx, method_cols, nrc_cols, id_cols)
}
\seealso{
\code{?prep_emotions_data}

Other DiccionariosEmocion: 
\code{\link{analisis_componentes_general}()},
\code{\link{analizar_coherencia_metodos}()},
\code{\link{build_matrix_from_artifact}()},
\code{\link{compute_correlations}()},
\code{\link{compute_kappa}()},
\code{\link{cronbach_alpha}()},
\code{\link{evaluar_con_artefacto}()},
\code{\link{features_canonicas}()},
\code{\link{guardar_artefacto_modelos}()},
\code{\link{normalizar_emociones}()},
\code{\link{predecir_con_artefacto}()},
\code{\link{preparar_features_tokens}()}
}
\concept{DiccionariosEmocion}
