% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_pca_cluster.R
\name{resModeloCA}
\alias{resModeloCA}
\title{resModeloCA}
\usage{
resModeloCA(dataCA, titulo = NULL, col_paleta = NULL)
}
\arguments{
\item{dataCA}{Matriz o data.frame numerico (tabla de contingencia).}

\item{titulo}{(Opcional) titulo descriptivo.}

\item{col_paleta}{(Opcional) paleta para los graficos (vector de colores o nombre).}
}
\value{
Lista con elementos: \code{Tabla}, \code{PorcentajeGlobal}, \code{PorcentajeFila},
\code{PorcentajeCol}, \code{ChiCuadrado}, \code{P_Valor}, \code{modelo},
\code{VarianzaExplicada}, \code{BiplotSimetrico}, \code{BiplotAsimetrico},
\code{PlotCos2Fila}, \code{PlotCos2Col}.
}
\description{
Ejecuta CA sobre una tabla de contingencia y retorna porcentajes,
chi-cuadrado, varianza explicada y biplots listos para graficar con \code{ggplot2}.
}
\details{
Analisis de Correspondencias (CA) con metricas y biplots
}
\examples{
\dontrun{
library(FactoMineR); library(factoextra); library(corrplot)
tb <- as.table(matrix(c(12,7,3, 5,6,4), nrow=2, byrow=TRUE))
res <- resModeloCA(tb, "Ejemplo CA")
res$BiplotSimetrico
}
}
\seealso{
Other modelado_supervisado: 
\code{\link{ajustar_modelo_polinomial}()},
\code{\link{ajustar_svm_multiclase}()},
\code{\link{entrenar_modelo_nivel}()},
\code{\link{entrenar_modelos_emocionales}()},
\code{\link{resModeloCluster}()},
\code{\link{resModeloMFA}()},
\code{\link{simular_svm_multiclase}()}
}
\concept{modelado_supervisado}
