% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_pca_cluster.R
\name{resModeloMFA}
\alias{resModeloMFA}
\title{resModeloMFA}
\usage{
resModeloMFA(dataMFA, grupos, tipos, nombres, rangoDim = NULL, indice)
}
\arguments{
\item{dataMFA}{Data.frame con bloques de variables.}

\item{grupos}{Vector de enteros con tamanos de cada bloque.}

\item{tipos}{Vector de tipos por bloque ("s", "n", "f", etc.).}

\item{nombres}{Nombres de grupos.}

\item{rangoDim}{Numero de grupos suplementarios (opcional).}

\item{indice}{Indice de columna para \code{habillage} en individuos.}
}
\value{
Lista con modelo, varianza y graficos.
}
\description{
Ejecuta MFA con FactoMineR devolviendo varianza, variables (cuantitativas/cualitativas)
y graficos listos de \code{factoextra}.
}
\details{
MFA (Multiple Factor Analysis) con visualizaciones clave
}
\examples{
\dontrun{
# Ver ?FactoMineR::MFA para un ejemplo detallado
}
}
\seealso{
Other modelado_supervisado: 
\code{\link{ajustar_modelo_polinomial}()},
\code{\link{ajustar_svm_multiclase}()},
\code{\link{entrenar_modelo_nivel}()},
\code{\link{entrenar_modelos_emocionales}()},
\code{\link{resModeloCA}()},
\code{\link{resModeloCluster}()},
\code{\link{simular_svm_multiclase}()}
}
\concept{modelado_supervisado}
