% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_pca_cluster.R
\name{resModeloPCA}
\alias{resModeloPCA}
\title{resModeloPCA}
\usage{
resModeloPCA(dataPCA, indCluster)
}
\arguments{
\item{dataPCA}{Data.frame o matriz donde una columna (indCluster) es factor/grupo.}

\item{indCluster}{Indice (entero, 1..p) de la columna que representa el grupo.}
}
\value{
Lista con modelo PCA, varianza, objetos de variables/individuos y graficos.
}
\description{
Ejecuta PCA (FactoMineR) y devuelve objetos graficables (factoextra)
y clustering k-means de variables activas.
}
\details{
PCA con graficos clave y clustering de variables
}
\examples{
\dontrun{
library(FactoMineR); library(factoextra); set.seed(1)
X <- iris[,1:4]; g <- iris$Species; df <- cbind(X, grupo = g)
res <- resModeloPCA(df, indCluster = 5)
res$graficaBiplot
}
}
\seealso{
Other reduccion_dimensional: 
\code{\link{ajustar_pca_svm}()},
\code{\link{estrategia_p4_pca_cluster}()},
\code{\link{generar_pca_emocional}()},
\code{\link{graficar_pca_emociones}()},
\code{\link{graficar_pca_emociones_etiquetas}()},
\code{\link{pca_tokens_biplot}()},
\code{\link{pca_tokens_plot}()},
\code{\link{run_pca}()},
\code{\link{simular_pca_svm}()}
}
\concept{reduccion_dimensional}
