% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_emociones.R
\name{run_pca}
\alias{run_pca}
\title{PCA (prcomp) con salida ordenada}
\usage{
run_pca(df, cols, scale = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{cols}{columnas numericas a incluir}

\item{scale}{TRUE para escalar internamente (recomendado si no pasaste z-score)}
}
\value{
list con objeto prcomp, var_exp, scores, loadings
}
\description{
Descripción (auto). Completar con propósito, entradas y salidas.
}
\seealso{
Other reduccion_dimensional: 
\code{\link{ajustar_pca_svm}()},
\code{\link{estrategia_p4_pca_cluster}()},
\code{\link{generar_pca_emocional}()},
\code{\link{graficar_pca_emociones}()},
\code{\link{graficar_pca_emociones_etiquetas}()},
\code{\link{pca_tokens_biplot}()},
\code{\link{pca_tokens_plot}()},
\code{\link{resModeloPCA}()},
\code{\link{simular_pca_svm}()}
}
\concept{reduccion_dimensional}
