% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_emociones.R
\name{save_plot_list}
\alias{save_plot_list}
\title{Compacta lista de gr치ficos (si existen) en disco con prefijo}
\usage{
save_plot_list(gl, out_dir, prefix)
}
\arguments{
\item{gl}{lista de ggplots}

\item{out_dir}{carpeta}

\item{prefix}{prefijo de archivo}
}
\value{
(auto) Verificar y completar la descripci칩n del valor retornado.
}
\description{
Descripci칩n (auto). Completar con prop칩sito, entradas y salidas.
}
\seealso{
Other visualizacion: 
\code{\link{CrearGraficoSentimiento}()},
\code{\link{estrategia_p6_estabilidad_visual}()},
\code{\link{graficar_emociones_agrupadas}()},
\code{\link{graficar_importancia_variables}()},
\code{\link{graficar_pr_multiclase}()},
\code{\link{graficar_roc_multiclase}()},
\code{\link{visualizar_curvas_sentimiento}()},
\code{\link{visualizar_mapa_emocional}()},
\code{\link{visualizar_modelo_polinomial}()}
}
\concept{visualizacion}
