% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos_aprendizaje.R
\name{seleccionar_features_tokens}
\alias{seleccionar_features_tokens}
\title{Seleccionar columnas numericas relevantes por defecto para tokens}
\usage{
seleccionar_features_tokens(
  tokens,
  extra = c("valencia_plus", "valencia_minus", "activacion")
)
}
\arguments{
\item{tokens}{data.frame con columnas numericas de metodos y derivadas}

\item{extra}{vector opcional con nombres adicionales}

\item{"valencia_minus"}{(auto) TODO: describir par치metro.}

\item{"activacion"}{(auto) TODO: describir par치metro.}
}
\value{
vector de nombres de columnas numericas a usar
}
\description{
Descripci칩n (auto). Completar con prop칩sito, entradas y salidas.
}
