% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_modelos.R
\name{simular_svm_multiclase}
\alias{simular_svm_multiclase}
\title{Simulacion usando SVM Multiclase}
\usage{
simular_svm_multiclase(
  modelo,
  nuevos_datos,
  columnas_referencia = NULL,
  valores_reales = NULL,
  mostrar_grafico = FALSE
)
}
\arguments{
\item{modelo}{objeto de clase \code{svm} entrenado (salida de \code{ajustar_svm_multiclase()$modelo})}

\item{nuevos_datos}{data.frame con las columnas predictoras, en el mismo orden que en el entrenamiento}

\item{columnas_referencia}{vector opcional con las columnas que deben verificarse y reordenarse}

\item{valores_reales}{vector factor opcional con las clases verdaderas para calcular matriz de confusi칩n}

\item{mostrar_grafico}{l칩gico, si es TRUE se imprime una matriz de confusi칩n y gr치fico de calor con ggplot2}
}
\value{
una lista con \code{predicciones} y \code{probabilidades}
}
\description{
Genera predicciones a partir de un modelo SVM multiclase entrenado con e1071.
}
\examples{
modelo_svm <- ajustar_svm_multiclase(acdfx, columnas_predictoras = vPred, columna_clase = "categoria")
nuevos <- head(acdfx, 3)
resultado <- simular_svm_multiclase(modelo_svm$modelo, nuevos, columnas_referencia = vPred,
                                    valores_reales = nuevos$categoria, mostrar_grafico = TRUE)
}
\seealso{
Other modelado_supervisado: 
\code{\link{ajustar_modelo_polinomial}()},
\code{\link{ajustar_svm_multiclase}()},
\code{\link{entrenar_modelo_nivel}()},
\code{\link{entrenar_modelos_emocionales}()},
\code{\link{resModeloCA}()},
\code{\link{resModeloCluster}()},
\code{\link{resModeloMFA}()}
}
\concept{modelado_supervisado}
